handle error upload google cloud

+++ app/Http/Controllers/V1/MissionController.php
@@ -308,7 +308,9 @@ class MissionController extends Controller
 
         } catch (\exception $exception){
          
-            Storage::disk('gcs')->delete($storage->file_path.'/'.$storage->file_name);  
+            if(isset($storage) && isset($storage->file_path) && isset($storage->file_name))
+                Storage::disk('gcs')->delete($storage->file_path.'/'.$storage->file_name);  
+            
             DB::rollBack();
 
             return (new ResponseTransformer)->toJson(500,$exception->getMessage(),false);
